app/requirements.txt
asgiref==3.5.2
Django==3.2
django-autocomplete-light==3.9.4
django-debug-toolbar==3.4.0
django-postgres-extra==2.0.6
django-split-settings==1.1.0
flake8==4.0.0
gunicorn==20.1.0
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.6.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
zipp==3.8.0
